+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,54 +1,12 @@
 package com.techcourse.service;
 
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import nextstep.jdbc.exception.DataAccessException;
-import nextstep.jdbc.exception.EmptyResultDataAccessException;
-import nextstep.jdbc.exception.IncorrectResultSizeDataAccessException;
-import nextstep.jdbc.exception.JdbcConnectionException;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.TransactionStatus;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final PlatformTransactionManager transactionManager;
+    User findById(long id);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.transactionManager = new DataSourceTransactionManager(DataSourceConfig.getInstance());
-    }
+    void insert(User user);
 
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        final TransactionStatus transactionStatus = transactionManager.getTransaction(
-                new DefaultTransactionDefinition());
-        try {
-            final var user = findById(id);
-            user.changePassword(newPassword);
-            userDao.update(user);
-            userHistoryDao.log(new UserHistory(user, createBy));
-            transactionManager.commit(transactionStatus);
-        } catch (final DataAccessException |
-                       EmptyResultDataAccessException |
-                       IncorrectResultSizeDataAccessException |
-                       JdbcConnectionException exception) {
-            transactionManager.rollback(transactionStatus);
-            throw exception;
-        }
-    }
+    void changePassword(long id, String newPassword, String createBy);
 }
